+++ domain/src/main/java/com/pens/planduit/domain/models/request/IncomeZakatRequest.kt
@@ -3,8 +3,8 @@ package com.pens.planduit.domain.models.request
 import com.google.gson.annotations.SerializedName
 
 data class IncomeZakatRequest(
-    val income: Int,
-    @SerializedName("another_income") val anotherIncome: Int,
-    val expenditure: Int,
+    val income: Long,
+    @SerializedName("another_income") val anotherIncome: Long,
+    val expenditure: Long,
     @SerializedName("time_type") val timeType: String
 )
