+++ domain/src/main/java/com/pens/planduit/domain/models/request/InvestmentRequest.kt
@@ -3,10 +3,10 @@ package com.pens.planduit.domain.models.request
 import com.google.gson.annotations.SerializedName
 
 data class InvestmentRequest(
-    @SerializedName("target_money") val targetMoney : Int,
+    @SerializedName("target_money") val targetMoney : Long,
     @SerializedName("target_time") val targetTime : Int,
-    @SerializedName("initial_money") val initialMoney : Int,
-    @SerializedName("money_investment") val moneyInvestment : Int,
+    @SerializedName("initial_money") val initialMoney : Long,
+    @SerializedName("money_investment") val moneyInvestment : Long,
     @SerializedName("time_type") val timeType : String,
     @SerializedName("interest") val interest : Int
     )
