+++ domain/src/main/java/com/pens/planduit/domain/models/request/TradingZakatRequest.kt
@@ -3,9 +3,9 @@ package com.pens.planduit.domain.models.request
 import com.google.gson.annotations.SerializedName
 
 data class TradingZakatRequest(
-    @SerializedName("rotated_capital") val rotatedCapital : Int,
-    @SerializedName("current_recievable") val currentRecievable : Int,
-    val profit : Int,
-    @SerializedName("current_payable") val currentPayable : Int,
-    val loss : Int
+    @SerializedName("rotated_capital") val rotatedCapital : Long,
+    @SerializedName("current_recievable") val currentRecievable : Long,
+    val profit : Long,
+    @SerializedName("current_payable") val currentPayable : Long,
+    val loss : Long
 )
